code cleanup and populating dictLocs and dv offsets
