+++ merge.go
@@ -25,7 +25,6 @@ import (
 
 	"github.com/RoaringBitmap/roaring"
 	seg "github.com/blevesearch/scorch_segment_api/v2"
-	"github.com/blevesearch/vellum"
 	"github.com/golang/snappy"
 )
 
@@ -142,11 +141,6 @@ func MergeToWriter(segments []*SegmentBase, drops []*roaring.Bitmap,
 			return nil, 0, 0, 0, 0, nil, nil, nil, 0, err
 		}
 
-		// dictLocs, docValueOffset, err = persistMergedRest(segments, drops, fieldsInv, fieldsMap, fieldsSame, newDocNums, numDocs, chunkMode, cr, closeCh)
-		// if err != nil {
-		// 	return nil, 0, 0, 0, 0, nil, nil, nil, 0, err
-		// }
-
 		// at this point, ask each section implementation to merge itself
 		for i, x := range segmentSections {
 			mergeOpaque[int(i)] = x.InitOpaque(args)
@@ -196,329 +190,6 @@ func computeNewDocCount(segments []*SegmentBase, drops []*roaring.Bitmap) uint64
 		}
 	}
 	return newDocCount
-}
-
-func mergeAndPersistInvertedSection(segments []*SegmentBase, dropsIn []*roaring.Bitmap,
-	fieldsInv []string, fieldsMap map[string]uint16, fieldsSame bool,
-	newDocNumsIn [][]uint64, newSegDocCount uint64, chunkMode uint32,
-	w *CountHashWriter, closeCh chan struct{}) (map[int]int, uint64, error) {
-	var bufMaxVarintLen64 []byte = make([]byte, binary.MaxVarintLen64)
-	var bufLoc []uint64
-
-	var postings *PostingsList
-	var postItr *PostingsIterator
-
-	fieldAddrs := make(map[int]int)
-	dictOffsets := make([]uint64, len(fieldsInv))
-	fieldDvLocsStart := make([]uint64, len(fieldsInv))
-	fieldDvLocsEnd := make([]uint64, len(fieldsInv))
-
-	// these int coders are initialized with chunk size 1024
-	// however this will be reset to the correct chunk size
-	// while processing each individual field-term section
-	tfEncoder := newChunkedIntCoder(1024, newSegDocCount-1)
-	locEncoder := newChunkedIntCoder(1024, newSegDocCount-1)
-
-	var vellumBuf bytes.Buffer
-	newVellum, err := vellum.New(&vellumBuf, nil)
-	if err != nil {
-		return nil, 0, err
-	}
-
-	newRoaring := roaring.NewBitmap()
-
-	// for each field
-	for fieldID, fieldName := range fieldsInv {
-		// collect FST iterators from all active segments for this field
-		var newDocNums [][]uint64
-		var drops []*roaring.Bitmap
-		var dicts []*Dictionary
-		var itrs []vellum.Iterator
-
-		var segmentsInFocus []*SegmentBase
-
-		for segmentI, segment := range segments {
-			// check for the closure in meantime
-			if isClosed(closeCh) {
-				return nil, 0, seg.ErrClosed
-			}
-
-			dict, err2 := segment.dictionary(fieldName)
-			if err2 != nil {
-				return nil, 0, err2
-			}
-			if dict != nil && dict.fst != nil {
-				itr, err2 := dict.fst.Iterator(nil, nil)
-				if err2 != nil && err2 != vellum.ErrIteratorDone {
-					return nil, 0, err2
-				}
-				if itr != nil {
-					newDocNums = append(newDocNums, newDocNumsIn[segmentI])
-					if dropsIn[segmentI] != nil && !dropsIn[segmentI].IsEmpty() {
-						drops = append(drops, dropsIn[segmentI])
-					} else {
-						drops = append(drops, nil)
-					}
-					dicts = append(dicts, dict)
-					itrs = append(itrs, itr)
-					segmentsInFocus = append(segmentsInFocus, segment)
-				}
-			}
-		}
-
-		var prevTerm []byte
-
-		newRoaring.Clear()
-
-		var lastDocNum, lastFreq, lastNorm uint64
-
-		// determines whether to use "1-hit" encoding optimization
-		// when a term appears in only 1 doc, with no loc info,
-		// has freq of 1, and the docNum fits into 31-bits
-		use1HitEncoding := func(termCardinality uint64) (bool, uint64, uint64) {
-			if termCardinality == uint64(1) && locEncoder.FinalSize() <= 0 {
-				docNum := uint64(newRoaring.Minimum())
-				if under32Bits(docNum) && docNum == lastDocNum && lastFreq == 1 {
-					return true, docNum, lastNorm
-				}
-			}
-			return false, 0, 0
-		}
-
-		finishTerm := func(term []byte) error {
-			tfEncoder.Close()
-			locEncoder.Close()
-
-			postingsOffset, err := writePostings(newRoaring,
-				tfEncoder, locEncoder, use1HitEncoding, w, bufMaxVarintLen64)
-			if err != nil {
-				return err
-			}
-
-			if postingsOffset > 0 {
-				err = newVellum.Insert(term, postingsOffset)
-				if err != nil {
-					return err
-				}
-			}
-
-			newRoaring.Clear()
-
-			tfEncoder.Reset()
-			locEncoder.Reset()
-
-			lastDocNum = 0
-			lastFreq = 0
-			lastNorm = 0
-
-			return nil
-		}
-
-		enumerator, err := newEnumerator(itrs)
-
-		for err == nil {
-			term, itrI, postingsOffset := enumerator.Current()
-
-			if !bytes.Equal(prevTerm, term) {
-				// check for the closure in meantime
-				if isClosed(closeCh) {
-					return nil, 0, seg.ErrClosed
-				}
-
-				// if the term changed, write out the info collected
-				// for the previous term
-				err = finishTerm(prevTerm)
-				if err != nil {
-					return nil, 0, err
-				}
-			}
-			if !bytes.Equal(prevTerm, term) || prevTerm == nil {
-				// compute cardinality of field-term in new seg
-				var newCard uint64
-				lowItrIdxs, lowItrVals := enumerator.GetLowIdxsAndValues()
-				for i, idx := range lowItrIdxs {
-					pl, err := dicts[idx].postingsListFromOffset(lowItrVals[i], drops[idx], nil)
-					if err != nil {
-						return nil, 0, err
-					}
-					newCard += pl.Count()
-				}
-				// compute correct chunk size with this
-				chunkSize, err := getChunkSize(chunkMode, newCard, newSegDocCount)
-				if err != nil {
-					return nil, 0, err
-				}
-				// update encoders chunk
-				tfEncoder.SetChunkSize(chunkSize, newSegDocCount-1)
-				locEncoder.SetChunkSize(chunkSize, newSegDocCount-1)
-			}
-
-			postings, err = dicts[itrI].postingsListFromOffset(
-				postingsOffset, drops[itrI], postings)
-			if err != nil {
-				return nil, 0, err
-			}
-
-			postItr = postings.iterator(true, true, true, postItr)
-
-			if fieldsSame {
-				// can optimize by copying freq/norm/loc bytes directly
-				lastDocNum, lastFreq, lastNorm, err = mergeTermFreqNormLocsByCopying(
-					term, postItr, newDocNums[itrI], newRoaring,
-					tfEncoder, locEncoder)
-			} else {
-				lastDocNum, lastFreq, lastNorm, bufLoc, err = mergeTermFreqNormLocs(
-					fieldsMap, term, postItr, newDocNums[itrI], newRoaring,
-					tfEncoder, locEncoder, bufLoc)
-			}
-			if err != nil {
-				return nil, 0, err
-			}
-
-			prevTerm = prevTerm[:0] // copy to prevTerm in case Next() reuses term mem
-			prevTerm = append(prevTerm, term...)
-
-			err = enumerator.Next()
-		}
-		if err != vellum.ErrIteratorDone {
-			return nil, 0, err
-		}
-
-		err = finishTerm(prevTerm)
-		if err != nil {
-			return nil, 0, err
-		}
-
-		dictOffset := uint64(w.Count())
-
-		err = newVellum.Close()
-		if err != nil {
-			return nil, 0, err
-		}
-		vellumData := vellumBuf.Bytes()
-
-		// write out the length of the vellum data
-		n := binary.PutUvarint(bufMaxVarintLen64, uint64(len(vellumData)))
-		_, err = w.Write(bufMaxVarintLen64[:n])
-		if err != nil {
-			return nil, 0, err
-		}
-
-		// write this vellum to disk
-		_, err = w.Write(vellumData)
-		if err != nil {
-			return nil, 0, err
-		}
-
-		dictOffsets[fieldID] = dictOffset
-
-		// perhaps rather than populating the slice we can just write the offsets at
-		// the beginning/end of the section. the format followed should be consistent
-		// across process, persist and merge
-		//
-		// get the field doc value offset (start)
-		fieldDvLocsStart[fieldID] = uint64(w.Count())
-
-		// update the field doc values
-		// NOTE: doc values continue to use legacy chunk mode
-		chunkSize, err := getChunkSize(LegacyChunkMode, 0, 0)
-		if err != nil {
-			return nil, 0, err
-		}
-		fdvEncoder := newChunkedContentCoder(chunkSize, newSegDocCount-1, w, true)
-
-		fdvReadersAvailable := false
-		var dvIterClone *docValueReader
-		for segmentI, segment := range segmentsInFocus {
-			// check for the closure in meantime
-			if isClosed(closeCh) {
-				return nil, 0, seg.ErrClosed
-			}
-
-			fieldIDPlus1 := uint16(segment.fieldsMap[fieldName])
-			if dvIter, exists := segment.fieldDvReaders[fieldIDPlus1-1]; exists &&
-				dvIter != nil {
-				fdvReadersAvailable = true
-				dvIterClone = dvIter.cloneInto(dvIterClone)
-				err = dvIterClone.iterateAllDocValues(segment, func(docNum uint64, terms []byte) error {
-					if newDocNums[segmentI][docNum] == docDropped {
-						return nil
-					}
-					err := fdvEncoder.Add(newDocNums[segmentI][docNum], terms)
-					if err != nil {
-						return err
-					}
-					return nil
-				})
-				if err != nil {
-					return nil, 0, err
-				}
-			}
-		}
-
-		if fdvReadersAvailable {
-			err = fdvEncoder.Close()
-			if err != nil {
-				return nil, 0, err
-			}
-
-			// persist the doc value details for this field
-			_, err = fdvEncoder.Write()
-			if err != nil {
-				return nil, 0, err
-			}
-
-			// get the field doc value offset (end)
-			fieldDvLocsEnd[fieldID] = uint64(w.Count())
-		} else {
-			fieldDvLocsStart[fieldID] = fieldNotUninverted
-			fieldDvLocsEnd[fieldID] = fieldNotUninverted
-		}
-
-		fieldStart := w.Count()
-
-		err = binary.Write(w, binary.BigEndian, fieldDvLocsStart[fieldID])
-		if err != nil {
-			return nil, 0, err
-		}
-
-		err = binary.Write(w, binary.BigEndian, fieldDvLocsEnd[fieldID])
-		if err != nil {
-			return nil, 0, err
-		}
-
-		err = binary.Write(w, binary.BigEndian, dictOffsets[fieldID])
-		if err != nil {
-			return nil, 0, err
-		}
-
-		fieldAddrs[fieldID] = fieldStart
-
-		// reset vellum buffer and vellum builder
-		vellumBuf.Reset()
-		err = newVellum.Reset(&vellumBuf)
-		if err != nil {
-			return nil, 0, err
-		}
-	}
-
-	fieldDvLocsOffset := uint64(w.Count())
-
-	// buf := bufMaxVarintLen64
-	// for i := 0; i < len(fieldDvLocsStart); i++ {
-	// 	n := binary.PutUvarint(buf, fieldDvLocsStart[i])
-	// 	_, err := w.Write(buf[:n])
-	// 	if err != nil {
-	// 		return nil, 0, err
-	// 	}
-	// 	n = binary.PutUvarint(buf, fieldDvLocsEnd[i])
-	// 	_, err = w.Write(buf[:n])
-	// 	if err != nil {
-	// 		return nil, 0, err
-	// 	}
-	// }
-
-	return fieldAddrs, fieldDvLocsOffset, nil
 }
 
 func mergeTermFreqNormLocsByCopying(term []byte, postItr *PostingsIterator,
