+++ new.go
@@ -17,7 +17,6 @@ package zap
 import (
 	"bytes"
 	"encoding/binary"
-	"math"
 	"sort"
 	"sync"
 	"sync/atomic"
@@ -297,15 +296,6 @@ func (s *interim) convert() (uint64, uint64, uint64, []uint64, uint64, error) {
 		s.opaque[int(i)] = x.InitOpaque(args)
 	}
 
-	// this very specific to inverted index, better to just invoke this function within
-	// 1/index. the main point of this function is init i believe, so the InitData
-	// should serve the purpose.
-	// s.prepareDicts()
-
-	// for _, dict := range s.DictKeys {
-	// 	sort.Strings(dict)
-	// }
-
 	s.processDocuments()
 
 	storedIndexOffset, err := s.writeStoredFields()
@@ -316,20 +306,8 @@ func (s *interim) convert() (uint64, uint64, uint64, []uint64, uint64, error) {
 	var fdvIndexOffset uint64
 	var dictOffsets []uint64
 
-	// the writeDicts is something specific to inverted index stuff.
-	// it writes the postings list, doc values and term dictionary
-	// to the writer. so, these can be made part of the section.Persist()
-	// API.
-	// if len(s.results) > 0 {
-	// 	fdvIndexOffset, dictOffsets, err = s.writeDicts()
-	// 	if err != nil {
-	// 		return 0, 0, 0, nil, 0, err
-	// 	}
-	// } else {
-	// 	dictOffsets = make([]uint64, len(s.FieldsInv))
-	// }
-
-	// we can persist other indexes at this point
+	// we can persist the various sections at this point.
+	// the rule of thumb here is that each section must persist field wise.
 	for _, x := range segmentSections {
 		_, err = x.Persist(s.opaque, s.w)
 		if err != nil {
@@ -371,107 +349,6 @@ func (s *interim) getOrDefineField(fieldName string) int {
 	}
 
 	return int(fieldIDPlus1 - 1)
-}
-
-// fill Dicts and DictKeys from analysis results
-func (s *interim) prepareDicts() {
-	var pidNext int
-
-	var totTFs int
-	var totLocs int
-
-	visitField := func(field index.Field) {
-		fieldID := uint16(s.getOrDefineField(field.Name()))
-
-		dict := s.Dicts[fieldID]
-		dictKeys := s.DictKeys[fieldID]
-		tfs := field.AnalyzedTokenFrequencies()
-		for term, tf := range tfs {
-			pidPlus1, exists := dict[term]
-			if !exists {
-				pidNext++
-				pidPlus1 = uint64(pidNext)
-
-				dict[term] = pidPlus1
-				dictKeys = append(dictKeys, term)
-
-				s.numTermsPerPostingsList = append(s.numTermsPerPostingsList, 0)
-				s.numLocsPerPostingsList = append(s.numLocsPerPostingsList, 0)
-			}
-
-			pid := pidPlus1 - 1
-
-			s.numTermsPerPostingsList[pid] += 1
-			s.numLocsPerPostingsList[pid] += len(tf.Locations)
-
-			totLocs += len(tf.Locations)
-		}
-
-		totTFs += len(tfs)
-
-		s.DictKeys[fieldID] = dictKeys
-	}
-
-	for _, result := range s.results {
-		// walk each composite field
-		result.VisitComposite(func(field index.CompositeField) {
-			visitField(field)
-		})
-
-		// walk each field
-		result.VisitFields(visitField)
-	}
-
-	numPostingsLists := pidNext
-
-	if cap(s.Postings) >= numPostingsLists {
-		s.Postings = s.Postings[:numPostingsLists]
-	} else {
-		postings := make([]*roaring.Bitmap, numPostingsLists)
-		copy(postings, s.Postings[:cap(s.Postings)])
-		for i := 0; i < numPostingsLists; i++ {
-			if postings[i] == nil {
-				postings[i] = roaring.New()
-			}
-		}
-		s.Postings = postings
-	}
-
-	if cap(s.FreqNorms) >= numPostingsLists {
-		s.FreqNorms = s.FreqNorms[:numPostingsLists]
-	} else {
-		s.FreqNorms = make([][]interimFreqNorm, numPostingsLists)
-	}
-
-	if cap(s.freqNormsBacking) >= totTFs {
-		s.freqNormsBacking = s.freqNormsBacking[:totTFs]
-	} else {
-		s.freqNormsBacking = make([]interimFreqNorm, totTFs)
-	}
-
-	freqNormsBacking := s.freqNormsBacking
-	for pid, numTerms := range s.numTermsPerPostingsList {
-		s.FreqNorms[pid] = freqNormsBacking[0:0]
-		freqNormsBacking = freqNormsBacking[numTerms:]
-	}
-
-	if cap(s.Locs) >= numPostingsLists {
-		s.Locs = s.Locs[:numPostingsLists]
-	} else {
-		s.Locs = make([][]interimLoc, numPostingsLists)
-	}
-
-	if cap(s.locsBacking) >= totLocs {
-		s.locsBacking = s.locsBacking[:totLocs]
-	} else {
-		s.locsBacking = make([]interimLoc, totLocs)
-	}
-
-	locsBacking := s.locsBacking
-	for pid, numLocs := range s.numLocsPerPostingsList {
-		s.Locs[pid] = locsBacking[0:0]
-		locsBacking = locsBacking[numLocs:]
-	}
 }
 
 func (s *interim) processDocuments() {
@@ -694,226 +571,6 @@ func (s *interim) writeStoredFields() (
 
 func (s *interim) setBytesWritten(val uint64) {
 	atomic.StoreUint64(&s.bytesWritten, val)
-}
-
-func (s *interim) writeDicts() (fdvIndexOffset uint64, dictOffsets []uint64, err error) {
-	dictOffsets = make([]uint64, len(s.FieldsInv))
-
-	fdvOffsetsStart := make([]uint64, len(s.FieldsInv))
-	fdvOffsetsEnd := make([]uint64, len(s.FieldsInv))
-
-	buf := s.grabBuf(binary.MaxVarintLen64)
-
-	// these int coders are initialized with chunk size 1024
-	// however this will be reset to the correct chunk size
-	// while processing each individual field-term section
-	tfEncoder := newChunkedIntCoder(1024, uint64(len(s.results)-1))
-	locEncoder := newChunkedIntCoder(1024, uint64(len(s.results)-1))
-
-	var docTermMap [][]byte
-
-	if s.builder == nil {
-		s.builder, err = vellum.New(&s.builderBuf, nil)
-		if err != nil {
-			return 0, nil, err
-		}
-	}
-
-	for fieldID, terms := range s.DictKeys {
-		if cap(docTermMap) < len(s.results) {
-			docTermMap = make([][]byte, len(s.results))
-		} else {
-			docTermMap = docTermMap[0:len(s.results)]
-			for docNum := range docTermMap { // reset the docTermMap
-				docTermMap[docNum] = docTermMap[docNum][:0]
-			}
-		}
-
-		dict := s.Dicts[fieldID]
-
-		for _, term := range terms { // terms are already sorted
-			pid := dict[term] - 1
-
-			postingsBS := s.Postings[pid]
-
-			freqNorms := s.FreqNorms[pid]
-			freqNormOffset := 0
-
-			locs := s.Locs[pid]
-			locOffset := 0
-
-			chunkSize, err := getChunkSize(s.chunkMode, postingsBS.GetCardinality(), uint64(len(s.results)))
-			if err != nil {
-				return 0, nil, err
-			}
-			tfEncoder.SetChunkSize(chunkSize, uint64(len(s.results)-1))
-			locEncoder.SetChunkSize(chunkSize, uint64(len(s.results)-1))
-
-			postingsItr := postingsBS.Iterator()
-			for postingsItr.HasNext() {
-				docNum := uint64(postingsItr.Next())
-
-				freqNorm := freqNorms[freqNormOffset]
-
-				// check if freq/norm is enabled
-				if freqNorm.freq > 0 {
-					err = tfEncoder.Add(docNum,
-						encodeFreqHasLocs(freqNorm.freq, freqNorm.numLocs > 0),
-						uint64(math.Float32bits(freqNorm.norm)))
-				} else {
-					// if disabled, then skip the norm part
-					err = tfEncoder.Add(docNum,
-						encodeFreqHasLocs(freqNorm.freq, freqNorm.numLocs > 0))
-				}
-				if err != nil {
-					return 0, nil, err
-				}
-
-				if freqNorm.numLocs > 0 {
-					numBytesLocs := 0
-					for _, loc := range locs[locOffset : locOffset+freqNorm.numLocs] {
-						numBytesLocs += totalUvarintBytes(
-							uint64(loc.fieldID), loc.pos, loc.start, loc.end,
-							uint64(len(loc.arrayposs)), loc.arrayposs)
-					}
-
-					err = locEncoder.Add(docNum, uint64(numBytesLocs))
-					if err != nil {
-						return 0, nil, err
-					}
-					for _, loc := range locs[locOffset : locOffset+freqNorm.numLocs] {
-						err = locEncoder.Add(docNum,
-							uint64(loc.fieldID), loc.pos, loc.start, loc.end,
-							uint64(len(loc.arrayposs)))
-						if err != nil {
-							return 0, nil, err
-						}
-
-						err = locEncoder.Add(docNum, loc.arrayposs...)
-						if err != nil {
-							return 0, nil, err
-						}
-					}
-					locOffset += freqNorm.numLocs
-				}
-
-				freqNormOffset++
-
-				docTermMap[docNum] = append(
-					append(docTermMap[docNum], term...),
-					termSeparator)
-			}
-
-			tfEncoder.Close()
-			locEncoder.Close()
-			s.incrementBytesWritten(locEncoder.getBytesWritten())
-
-			postingsOffset, err :=
-				writePostings(postingsBS, tfEncoder, locEncoder, nil, s.w, buf)
-			if err != nil {
-				return 0, nil, err
-			}
-
-			if postingsOffset > uint64(0) {
-				err = s.builder.Insert([]byte(term), postingsOffset)
-				if err != nil {
-					return 0, nil, err
-				}
-			}
-
-			tfEncoder.Reset()
-			locEncoder.Reset()
-		}
-
-		err = s.builder.Close()
-		if err != nil {
-			return 0, nil, err
-		}
-
-		// record where this dictionary starts
-		dictOffsets[fieldID] = uint64(s.w.Count())
-
-		vellumData := s.builderBuf.Bytes()
-
-		// write out the length of the vellum data
-		n := binary.PutUvarint(buf, uint64(len(vellumData)))
-		_, err = s.w.Write(buf[:n])
-		if err != nil {
-			return 0, nil, err
-		}
-
-		// write this vellum to disk
-		_, err = s.w.Write(vellumData)
-		if err != nil {
-			return 0, nil, err
-		}
-
-		s.incrementBytesWritten(uint64(len(vellumData)))
-
-		// reset vellum for reuse
-		s.builderBuf.Reset()
-
-		err = s.builder.Reset(&s.builderBuf)
-		if err != nil {
-			return 0, nil, err
-		}
-
-		// write the field doc values
-		// NOTE: doc values continue to use legacy chunk mode
-		chunkSize, err := getChunkSize(LegacyChunkMode, 0, 0)
-		if err != nil {
-			return 0, nil, err
-		}
-
-		fdvEncoder := newChunkedContentCoder(chunkSize, uint64(len(s.results)-1), s.w, false)
-		if s.IncludeDocValues[fieldID] {
-			for docNum, docTerms := range docTermMap {
-				if len(docTerms) > 0 {
-					err = fdvEncoder.Add(uint64(docNum), docTerms)
-					if err != nil {
-						return 0, nil, err
-					}
-				}
-			}
-			err = fdvEncoder.Close()
-			if err != nil {
-				return 0, nil, err
-			}
-
-			s.incrementBytesWritten(fdvEncoder.getBytesWritten())
-
-			fdvOffsetsStart[fieldID] = uint64(s.w.Count())
-
-			_, err = fdvEncoder.Write()
-			if err != nil {
-				return 0, nil, err
-			}
-
-			fdvOffsetsEnd[fieldID] = uint64(s.w.Count())
-
-			fdvEncoder.Reset()
-		} else {
-			fdvOffsetsStart[fieldID] = fieldNotUninverted
-			fdvOffsetsEnd[fieldID] = fieldNotUninverted
-		}
-	}
-
-	fdvIndexOffset = uint64(s.w.Count())
-
-	for i := 0; i < len(fdvOffsetsStart); i++ {
-		n := binary.PutUvarint(buf, fdvOffsetsStart[i])
-		_, err := s.w.Write(buf[:n])
-		if err != nil {
-			return 0, nil, err
-		}
-		n = binary.PutUvarint(buf, fdvOffsetsEnd[i])
-		_, err = s.w.Write(buf[:n])
-		if err != nil {
-			return 0, nil, err
-		}
-	}
-
-	return fdvIndexOffset, dictOffsets, nil
 }
 
 // returns the total # of bytes needed to encode the given uint64's
